main/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-adk>=1.0.0
google-genai>=1.0.0
pydantic>=2.10.0
httpx>=0.28.0
supabase>=2.10.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
Pillow>=11.0.0
cachetools>=5.5.0
